# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface VoiceTag {
  name: string;
  type: string;
  aliases: string[];
  description: string;
  strategy: string;
  spoken: boolean;
  separateAudioEvent: boolean;
  postProcess: boolean;
}

export interface ParseVoiceTextRequest {
  /**
     * @minLength 1
     * @maxLength 20000
     */
  text: string;
}

export type ParsedVoiceTextEventsItem = { [key: string]: unknown };

export type ParsedVoiceTextTagsItem = {
  raw: string;
  normalized: string;
};

export interface ParsedVoiceText {
  events: ParsedVoiceTextEventsItem[];
  tags: ParsedVoiceTextTagsItem[];
}

export interface VoiceSummary {
  id: string;
  name: string;
  referenceText?: string;
  createdAt: string;
}

export type GenerateVoiceRequestSettings = { [key: string]: unknown };

export type GenerateVoiceRequest = ParseVoiceTextRequest & {
  settings?: GenerateVoiceRequestSettings;
};

export interface GenerationAccepted {
  generationId: string;
  status: string;
  /**
     * @minimum 0
     * @maximum 100
     */
  progress: number;
}

export interface GenerationStatus {
  generationId: string;
  status: string;
  /**
     * @minimum 0
     * @maximum 100
     */
  progress: number;
  audioAvailable: boolean;
  error?: string;
}

export type ListVoiceTags200 = {
  tags: VoiceTag[];
};

export type ListVoices200 = {
  voices: VoiceSummary[];
};

export type TranscribeVoiceAudio200 = {
  text: string;
};