# lib/api-zod/src/generated/types/parsedVoiceText.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ParsedVoiceTextEventsItem } from './parsedVoiceTextEventsItem';
import type { ParsedVoiceTextTagsItem } from './parsedVoiceTextTagsItem';

export interface ParsedVoiceText {
  events: ParsedVoiceTextEventsItem[];
  tags: ParsedVoiceTextTagsItem[];
}